fix: change constants and type location gf-12
