--- apps/frontend/src/assets/css/main.css
@@ -1,4 +1,0 @@
-main {
-	display: flex;
-	flex: 1;
-}
