+++ apps/frontend/src/assets/css/styles.css
@@ -1,5 +1,4 @@
 @import url("./fonts.css");
 @import url("./scaffolding.css");
 @import url("./media.css");
-@import url("./main.css");
 @import url("./variables.css");
