+++ apps/frontend/src/libs/types/navigation-item.type.ts
@@ -3,7 +3,7 @@ import { type AppRoute } from "~/libs/enums/app-route.enum.js";
 import { type ValueOf } from "~/libs/types/types.js";
 
 type NavigationItem = {
-	href: "#" | ValueOf<typeof AppRoute>;
+	href: ValueOf<typeof AppRoute>;
 	icon: IconName;
 	label: string;
 };
