+++ apps/frontend/src/libs/types/types.ts
@@ -1,4 +1,5 @@
 export { type AsyncThunkConfig } from "./async-thunk-config.type.js";
+export { type NavigationItem } from "./navigation-item.type.js";
 export {
 	type ServerErrorDetail,
 	type ServerErrorResponse,
